[refactor] use test resource
