+++ src/test/java/toy/bookchat/bookchat/security/token/TokenServiceTest.java
@@ -19,7 +19,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import toy.bookchat.bookchat.config.token.JwtTokenProperties;
 import toy.bookchat.bookchat.domain.user.ReadingTaste;
 import toy.bookchat.bookchat.domain.user.User;
 import toy.bookchat.bookchat.domain.user.api.dto.Token;
@@ -38,8 +37,6 @@ class TokenServiceTest {
     JwtTokenProvider jwtTokenProvider;
     @Spy
     JwtTokenManager jwtTokenManager;
-    @Mock
-    JwtTokenProperties jwtTokenProperties;
     @Mock
     UserRepository userRepository;
     @Mock
