+++ src/test/java/toy/bookchat/bookchat/security/token/jwt/JwtTokenProviderTest.java
@@ -21,7 +21,6 @@ class JwtTokenProviderTest {
 
     @Mock
     JwtTokenProperties jwtTokenProperties;
-
     @InjectMocks
     JwtTokenManagerImpl jwtTokenManager;
     @InjectMocks
