+++ src/test/java/toy/bookchat/bookchat/security/token/jwt/RefreshTokenRepositoryTest.java
@@ -7,10 +7,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import toy.bookchat.bookchat.config.query.JpaAuditingConfig;
-import toy.bookchat.bookchat.domain.configuration.TestConfig;
+import toy.bookchat.bookchat.domain.RepositoryTestConfiguration;
 
 @DataJpaTest
-@Import({JpaAuditingConfig.class, TestConfig.class})
+@Import({JpaAuditingConfig.class, RepositoryTestConfiguration.class})
 class RefreshTokenRepositoryTest {
 
     @Autowired
